+++ src/Agui/Widgets/TextBox/TextBox.cpp
@@ -588,7 +588,7 @@ namespace agui {
 		size_t letterLength = 0;
 		size_t ind = 0;
 
-		for(int i = 0; i < len; ++i)
+		for(size_t i = 0; i < len; ++i)
 		{
 			
 			//get the unicode character
